requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
Django==4.2.14
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
flake8==7.1.0
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pycodestyle==2.12.0
pyflakes==3.2.0
PyJWT==2.8.0
sqlparse==0.5.1
tomli==2.0.1
typing-extensions==4.12.2
